SSE-2165: Organise routes and controllers by app sections
Ensure each router and controller only touches one section of the app - each should have a single responsibility.
Add routers for service and base routes, and move those routes out of the other routers.
Move handlers around to ensure each router only imports handlers from one controller.

Move non-exported function to the bottom.
